requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
azure-ai-textanalytics==5.3.0
azure-common==1.1.28
azure-core==1.32.0
certifi==2025.1.31
charset-normalizer==3.4.1
distro==1.9.0
Django==5.1.6
djangorestframework==3.15.2
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
isodate==0.7.2
jiter==0.8.2
openai==1.61.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
